get_plane() changed.
Removed import of math, inverted signed distance seem to do same trick as previous angle. Shorter and clearer so changed.

Noticed calls to pymel.node.name() instead of nodeName() which caused issues at some cases so fixed this too.

Also removed variable definition of chain prior to my if/else - originally wrote this in Vim but had pycharm now tip me it was not needed.

Not sure about :rtype - but if it is good thing to add I'll make sure to add it everywhere.